[feat #184] S3 Presigned URL 업로드 로직 -> File 공통 도메인 하위로 이동
