+++ momentum-dao-be/src/main/java/com/dao/momentum/common/service/S3Service.java
@@ -1,6 +1,6 @@
 package com.dao.momentum.common.service;
 
-import com.dao.momentum.announcement.command.application.dto.response.FilePresignedUrlResponse;
+import com.dao.momentum.file.command.application.dto.response.FilePresignedUrlResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.http.client.utils.URIBuilder;
