+++ momentum-dao-be/src/main/java/com/dao/momentum/file/command/application/dto/request/FilePresignedUrlRequest.java
@@ -1,3 +1,3 @@
-package com.dao.momentum.announcement.command.application.dto.request;
+package com.dao.momentum.file.command.application.dto.request;
 
 public record FilePresignedUrlRequest(String fileName, long sizeInBytes, String contentType) {}
