+++ momentum-dao-be/src/main/java/com/dao/momentum/file/command/application/dto/response/FilePresignedUrlResponse.java
@@ -1,3 +1,3 @@
-package com.dao.momentum.announcement.command.application.dto.response;
+package com.dao.momentum.file.command.application.dto.response;
 
 public record FilePresignedUrlResponse(String presignedUrl, String s3Key) {}
