+++ momentum-dao-be/src/test/java/com/dao/momentum/announcement/command/application/service/AnnouncementCommandServiceTest.java
@@ -2,19 +2,17 @@ package com.dao.momentum.announcement.command.application.service;
 
 import com.dao.momentum.announcement.command.application.dto.request.AnnouncementCreateRequest;
 import com.dao.momentum.announcement.command.application.dto.request.AnnouncementModifyRequest;
-import com.dao.momentum.announcement.command.application.dto.request.AttachmentRequest;
-import com.dao.momentum.announcement.command.application.dto.request.FilePresignedUrlRequest;
+import com.dao.momentum.file.command.application.dto.request.AttachmentRequest;
 import com.dao.momentum.announcement.command.application.dto.response.AnnouncementCreateResponse;
 import com.dao.momentum.announcement.command.application.dto.response.AnnouncementModifyResponse;
-import com.dao.momentum.announcement.command.application.dto.response.FilePresignedUrlResponse;
 import com.dao.momentum.announcement.command.application.mapper.AnnouncementMapper;
 import com.dao.momentum.announcement.command.domain.aggregate.Announcement;
-import com.dao.momentum.announcement.command.domain.aggregate.File;
 import com.dao.momentum.announcement.command.domain.repository.AnnouncementRepository;
-import com.dao.momentum.announcement.command.domain.repository.FileRepository;
 import com.dao.momentum.announcement.exception.AnnouncementAccessDeniedException;
 import com.dao.momentum.announcement.exception.NoSuchAnnouncementException;
 import com.dao.momentum.common.service.S3Service;
+import com.dao.momentum.file.command.domain.aggregate.File;
+import com.dao.momentum.file.command.domain.repository.FileRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -47,50 +45,6 @@ class AnnouncementCommandServiceTest {
 
     @InjectMocks
     private AnnouncementCommandService announcementCommandService;
-
-    @Test
-    @DisplayName("Presigned URL 생성 성공")
-    void generatePresignedUrl_success() {
-        // given
-        FilePresignedUrlRequest request = new FilePresignedUrlRequest("test.png", 1024 * 1024, "image/png");
-
-        when(s3Service.extractFileExtension("test.png")).thenReturn("png");
-        when(s3Service.sanitizeFilename("test.png")).thenReturn("test.png");
-        when(s3Service.generatePresignedUploadUrlWithKey(anyString(), eq("image/png")))
-                .thenReturn(new FilePresignedUrlResponse("https://presigned.url", "announcements/uuid/test.png"));
-
-        // when
-        FilePresignedUrlResponse response = announcementCommandService.generatePresignedUrl(request);
-
-        // then
-        assertNotNull(response);
-        assertTrue(response.presignedUrl().startsWith("https://presigned.url"));
-        assertTrue(response.s3Key().startsWith("announcements/"));
-    }
-
-    @Test
-    @DisplayName("Presigned URL 생성 실패 - 파일 크기 초과")
-    void generatePresignedUrl_fail_fileTooLarge() {
-        FilePresignedUrlRequest request = new FilePresignedUrlRequest("test.png", 11 * 1024 * 1024, "image/png");
-
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> announcementCommandService.generatePresignedUrl(request));
-
-        assertEquals("파일은 10MB 이하만 업로드 가능합니다.", exception.getMessage());
-    }
-
-    @Test
-    @DisplayName("Presigned URL 생성 실패 - 허용되지 않은 확장자")
-    void generatePresignedUrl_fail_invalidExtension() {
-        FilePresignedUrlRequest request = new FilePresignedUrlRequest("malware.exe", 1024 * 1024, "application/octet-stream");
-
-        when(s3Service.extractFileExtension("malware.exe")).thenReturn("exe");
-
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> announcementCommandService.generatePresignedUrl(request));
-
-        assertEquals("허용되지 않은 파일 확장자입니다.", exception.getMessage());
-    }
 
     @Test
     @DisplayName("공지사항 생성 성공")
